+++ packages/noteberry-editor/src/components/header/Navbar.tsx
@@ -1,0 +1,7 @@
+export function Navbar() {
+  return (
+    <>
+      <h2>Navbar</h2>
+    </>
+  );
+}