+++ packages/noteberry-editor/CHANGELOG.md
@@ -5,3 +5,4 @@ Follow the SemVer specification.
 ## 0.1.0
 
 - Initial release
+- Setup of a simple editor application with multiple pages to edit